Test /dev/tty implementation
--- Duck Review Questions & Answers ---

Q: What specific issue or requirement prompted the addition of "Final test with /dev/tty" to the file?
A: because mrinal said do

Q: How does this new line enhance the functionality or clarity of the existing tests?
A: it doesn't

Q: Are there any assumptions made about the environment or context in which this test will be executed that we should be aware of?
A: yes

Q: What potential side effects do you anticipate from including this new test with `/dev/tty` in the overall testing process?
A: all of them

Q: How do you plan to validate that the addition of this line is effective and does not introduce any regressions?
A: it will instroduce many regressions